@@ -2,7 +2,7 @@ package org.onedatashare.odsmetadata.entity;
 
 import lombok.Data;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.sql.Timestamp;
 import java.util.List;
 
